PARAMETRIC, CONSTRAINED CAD MODELS FOR THE WIN!
No really, re-implemented the bridge and head assembly in Fusion360 as parametric models. So awesome. Much Wow.

So this is the first STL dump of that work, which sets up the head to use an inexpensive D2826 BLDC and TR8x8 leadscrew on a 1:3.5 reduction driven from a NEMA17 mounted up top.